Renamed app package name to MediaPickerApp because of Manifest dex error
