+++ app/src/androidTest/java/com/coursion/mediapickerApp/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.coursion.mediapickerlib
+package com.coursion.mediapickerApp
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
