+++ app/src/main/java/com/coursion/mediapickerApp/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.coursion.mediapickerlib
+package com.coursion.mediapickerApp
 
 import android.Manifest
 import android.content.Intent
