+++ app/src/main/res/layout/content_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.coursion.mediapickerlib.MainActivity"
+    tools:context="com.coursion.mediapickerApp.MainActivity"
     tools:showIn="@layout/activity_main">
 
     <ImageView
