+++ mediapickerlib/src/main/java/com/coursion/freakycoder/mediapicker/fragments/VideoFragment.kt
@@ -15,8 +15,7 @@ import android.view.View
 import android.view.ViewGroup
 import com.coursion.freakycoder.mediapicker.adapters.BucketsAdapter
 import com.coursion.freakycoder.mediapicker.galleries.OpenGallery
-import com.coursion.mediapickerlib.R
-import kotlinx.android.synthetic.main.fragment_image.*
+import com.coursion.mediapickerLib.R
 
 import java.io.File
 import java.util.ArrayList
